Request #32: on-list status
NOTE: This is a proper FIX-COMMIT (previously only image files were
commited)

For status displaying icons were added to the file list. At the moment
there are 4 types of icons: (1)succeeded job (2)failed job (3)skipped by
user job and (4)new job. Perhaps icons for skipped and new jobs could be
changed because they might be a little confusing but choosing icons is
not a part of this request.